Update README.md
While the align="center" attribute was once a common method for centering content in HTML, it is now considered outdated in modern web development. Using this method restricts flexibility and does not provide full control over the layout of elements, especially when you need to customize the appearance across different screen sizes.

Signed-off-by: BARECHA ETANA <145106498+bar-echa@users.noreply.github.com>